fix getting value from max bot long pooling

+++ src/main/java/ru/akvine/wild/bot/services/integration/max/dto/response/LongPoolingSubscriptionResponse.java
@@ -12,5 +12,5 @@ public class LongPoolingSubscriptionResponse {
     private Update[] updates;
 
     @JsonProperty(value = "marker")
-    private int marker;
+    private long marker;
 }
